refactor(schemas): reorg constants
