+++ packages/schemas/src/types/index.ts
@@ -10,3 +10,4 @@ export * from './scope.js';
 export * from './role.js';
 export * from './verification-code.js';
 export * from './application.js';
+export * from './system.js';
